+++ src/main/java/fr/marcwrobel/jbanking/calendar/FinancialCalendars.java
@@ -106,10 +106,9 @@ public enum FinancialCalendars implements Calendar {
           PENTECOST_MONDAY,
           FEAST_OF_CORPUS_CHRISTI,
           // https://wikipedia.org/wiki/German_Unity_Day
-          new MonthDayHoliday(MonthDay.of(OCTOBER, 3)),
+          new MonthDayHoliday(OCTOBER, 3),
           // https://wikipedia.org/wiki/Reformation_Day - was a national holiday only on 2017
-          new YearRangeHoliday(
-              new MonthDayHoliday(MonthDay.of(OCTOBER, 31)), ValueRange.of(2017, 2017)),
+          new YearRangeHoliday(new MonthDayHoliday(OCTOBER, 31), 2017, 2017),
           CHRISTMAS_EVE,
           CHRISTMAS_DAY,
           SAINT_STEPHENS_DAY,
@@ -130,7 +129,7 @@ public enum FinancialCalendars implements Calendar {
           new ShiftedHoliday(NEW_YEAR_DAY, WEEKEND_TO_MONDAY),
           GOOD_FRIDAY,
           EASTER_MONDAY,
-          new MovedHoliday(Holidays.MAY_DAY, LocalDate.of(2020, 5, 4), LocalDate.of(2020, 5, 8)),
+          new MovedHoliday(Holidays.MAY_DAY, d(2020, 5, 4), d(2020, 5, 8)),
           // https://www.timeanddate.com/holidays/uk/spring-bank-holiday - suppression for jubilees
           new SuppressedHoliday(new DayOfWeekInMonthHoliday(-1, MONDAY, MAY), 2002, 2012, 2022),
           // https://www.timeanddate.com/holidays/uk/summer-bank-holiday
@@ -138,13 +137,13 @@ public enum FinancialCalendars implements Calendar {
           new ShiftedHoliday(CHRISTMAS_DAY, PLUS_TWO_DAYS),
           new ShiftedHoliday(BOXING_DAY, PLUS_TWO_DAYS),
           // golden jubilee
-          new FixedHoliday(LocalDate.of(2002, 6, 3), LocalDate.of(2002, 6, 4)),
+          new FixedHoliday(d(2002, 6, 3), d(2002, 6, 4)),
           // royal wedding
-          new FixedHoliday(LocalDate.of(2011, 4, 29)),
+          new FixedHoliday(d(2011, 4, 29)),
           // golden jubilee
-          new FixedHoliday(LocalDate.of(2012, 6, 4), LocalDate.of(2012, 6, 5)),
+          new FixedHoliday(d(2012, 6, 4), d(2012, 6, 5)),
           // platinum jubilee
-          new FixedHoliday(LocalDate.of(2022, 6, 2), LocalDate.of(2022, 6, 3)))),
+          new FixedHoliday(d(2022, 6, 2), d(2022, 6, 3)))),
 
   /**
    * Federal Reserve Bank of New York holidays.
@@ -166,7 +165,7 @@ public enum FinancialCalendars implements Calendar {
           // https://wikipedia.org/wiki/Memorial_Day
           new DayOfWeekInMonthHoliday(-1, MONDAY, MAY),
           // https://wikipedia.org/wiki/Independence_Day_(United_States)
-          new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(JULY, 4)), SUNDAY_TO_MONDAY),
+          new ShiftedHoliday(new MonthDayHoliday(JULY, 4), SUNDAY_TO_MONDAY),
           // https://wikipedia.org/wiki/Labor_Day
           new DayOfWeekInMonthHoliday(1, MONDAY, SEPTEMBER),
           // https://wikipedia.org/wiki/Columbus_Day
@@ -198,26 +197,22 @@ public enum FinancialCalendars implements Calendar {
           // https://wikipedia.org/wiki/Memorial_Day
           new DayOfWeekInMonthHoliday(-1, MONDAY, MAY),
           // https://wikipedia.org/wiki/Independence_Day_(United_States)
-          new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(JULY, 4)), CLOSEST_WEEKDAY),
+          new ShiftedHoliday(new MonthDayHoliday(JULY, 4), CLOSEST_WEEKDAY),
           // https://wikipedia.org/wiki/Labor_Day
           new DayOfWeekInMonthHoliday(1, MONDAY, SEPTEMBER),
           // https://wikipedia.org/wiki/Thanksgiving
           new DayOfWeekInMonthHoliday(4, THURSDAY, NOVEMBER),
           new ShiftedHoliday(CHRISTMAS_DAY, CLOSEST_WEEKDAY),
           // https://wikipedia.org/wiki/September_11_attacks
-          new FixedHoliday(
-              LocalDate.of(2001, 9, 11),
-              LocalDate.of(2001, 9, 12),
-              LocalDate.of(2001, 9, 13),
-              LocalDate.of(2001, 9, 14)),
+          new FixedHoliday(d(2001, 9, 11), d(2001, 9, 12), d(2001, 9, 13), d(2001, 9, 14)),
           // https://wikipedia.org/wiki/Death_and_state_funeral_of_Ronald_Reagan
-          new FixedHoliday(LocalDate.of(2004, 6, 11)),
+          new FixedHoliday(d(2004, 6, 11)),
           // https://wikipedia.org/wiki/Death_and_state_funeral_of_Gerald_Ford
-          new FixedHoliday(LocalDate.of(2007, 1, 2)),
+          new FixedHoliday(d(2007, 1, 2)),
           // https://wikipedia.org/wiki/Hurricane_Sandy
-          new FixedHoliday(LocalDate.of(2012, 10, 30)),
+          new FixedHoliday(d(2012, 10, 30)),
           // https://wikipedia.org/wiki/Death_and_state_funeral_of_George_H._W._Bush
-          new FixedHoliday(LocalDate.of(2018, 12, 5)))),
+          new FixedHoliday(d(2018, 12, 5)))),
 
   /**
    * Paris (France) financial district holidays.
@@ -243,7 +238,7 @@ public enum FinancialCalendars implements Calendar {
           // see http://jollyday.sourceforge.net/data/fr.html
           new SuppressedHoliday(PENTECOST_MONDAY, 2005, 2006, 2007),
           // https://wikipedia.org/wiki/Bastille_Day
-          new MonthDayHoliday(MonthDay.of(JULY, 14)),
+          new MonthDayHoliday(JULY, 14),
           ASSUMPTION_OF_MARY,
           ALL_SAINTS_DAY,
           ARMISTICE_DAY,
@@ -266,11 +261,11 @@ public enum FinancialCalendars implements Calendar {
           SUNDAY,
           new ShiftedHoliday(NEW_YEAR_DAY, WEEKEND_TO_MONDAY),
           // https://wikipedia.org/wiki/Australia_Day
-          new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(JANUARY, 26)), WEEKEND_TO_MONDAY),
+          new ShiftedHoliday(new MonthDayHoliday(JANUARY, 26), WEEKEND_TO_MONDAY),
           GOOD_FRIDAY,
           EASTER_MONDAY,
           // https://wikipedia.org/wiki/Anzac_Day
-          new MonthDayHoliday(MonthDay.of(APRIL, 25)),
+          new MonthDayHoliday(APRIL, 25),
           // https://wikipedia.org/wiki/Queen%27s_Official_Birthday
           new DayOfWeekInMonthHoliday(2, MONDAY, JUNE),
           // https://wikipedia.org/wiki/Bank_holiday
@@ -302,7 +297,7 @@ public enum FinancialCalendars implements Calendar {
           CHRISTMAS_DAY,
           SAINT_STEPHENS_DAY,
           // https://www.ecb.europa.eu/press/pr/date/2000/html/pr000525_2.en.html
-          new FixedHoliday(LocalDate.of(2001, 12, 31)))),
+          new FixedHoliday(d(2001, 12, 31)))),
 
   /**
    * Tokyo (Japan) financial district holidays.
@@ -318,12 +313,12 @@ public enum FinancialCalendars implements Calendar {
           SATURDAY,
           SUNDAY,
           new ShiftedHoliday(NEW_YEAR_DAY, SUNDAY_TO_MONDAY),
-          new MonthDayHoliday(MonthDay.of(JANUARY, 2)),
-          new MonthDayHoliday(MonthDay.of(JANUARY, 3)),
+          new MonthDayHoliday(JANUARY, 2),
+          new MonthDayHoliday(JANUARY, 3),
           // https://wikipedia.org/wiki/Coming_of_Age_Day
           new DayOfWeekInMonthHoliday(2, MONDAY, JANUARY),
           // https://wikipedia.org/wiki/National_Foundation_Day
-          new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(FEBRUARY, 11)), SUNDAY_TO_MONDAY),
+          new ShiftedHoliday(new MonthDayHoliday(FEBRUARY, 11), SUNDAY_TO_MONDAY),
           // https://www.timeanddate.com/holidays/japan/spring-equinox
           new ShiftedHoliday(
               new MovedHoliday(
@@ -349,49 +344,44 @@ public enum FinancialCalendars implements Calendar {
                   2030),
               SUNDAY_TO_MONDAY),
           // https://wikipedia.org/wiki/Sh%C5%8Dwa_Day
-          new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(APRIL, 29)), SUNDAY_TO_MONDAY),
+          new ShiftedHoliday(new MonthDayHoliday(APRIL, 29), SUNDAY_TO_MONDAY),
           // https://wikipedia.org/wiki/Constitution_Memorial_Day
           new YearRangeHoliday(
-              new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(MAY, 3)), SUNDAY_TO_MONDAY),
-              ValueRange.of(1950, 2006)),
+              new ShiftedHoliday(new MonthDayHoliday(MAY, 3), SUNDAY_TO_MONDAY), 1950, 2006),
           new YearRangeHoliday(
-              new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(MAY, 3)), SUNDAY_TO_WEDNESDAY),
-              ValueRange.of(2007, 2999)),
+              new ShiftedHoliday(new MonthDayHoliday(MAY, 3), SUNDAY_TO_WEDNESDAY), 2007, 2999),
           // https://wikipedia.org/wiki/Greenery_Day from 2007, holiday between holidays before
           new YearRangeHoliday(
-              new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(MAY, 4)), SUNDAY_TO_MONDAY),
-              ValueRange.of(1950, 2006)),
+              new ShiftedHoliday(new MonthDayHoliday(MAY, 4), SUNDAY_TO_MONDAY), 1950, 2006),
           new YearRangeHoliday(
-              new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(MAY, 4)), SUNDAY_TO_TUESDAY),
-              ValueRange.of(2007, 2999)),
+              new ShiftedHoliday(new MonthDayHoliday(MAY, 4), SUNDAY_TO_TUESDAY), 2007, 2999),
           // https://wikipedia.org/wiki/Children's_Day#Japan
-          new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(MAY, 5)), SUNDAY_TO_MONDAY),
+          new ShiftedHoliday(new MonthDayHoliday(MAY, 5), SUNDAY_TO_MONDAY),
           // https://en.wikipedia.org/wiki/Marine_Day
           new YearRangeHoliday(
-              new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(JULY, 20)), SUNDAY_TO_MONDAY),
-              ValueRange.of(1950, 2003)),
+              new ShiftedHoliday(new MonthDayHoliday(JULY, 20), SUNDAY_TO_MONDAY), 1950, 2003),
           new YearRangeHoliday(
               new MovedHoliday(
                   new DayOfWeekInMonthHoliday(3, MONDAY, JULY),
-                  LocalDate.of(2020, 7, 20), // The Olympics
-                  LocalDate.of(2020, 7, 23),
-                  LocalDate.of(2021, 7, 19), // Postponement of the Olympics
-                  LocalDate.of(2021, 7, 22)),
-              ValueRange.of(2003, 2999)),
+                  d(2020, 7, 20), // The Olympics
+                  d(2020, 7, 23),
+                  d(2021, 7, 19), // Postponement of the Olympics
+                  d(2021, 7, 22)),
+              2003,
+              2999),
           // https://wikipedia.org/wiki/Mountain_Day
           new YearRangeHoliday(
               new MovedHoliday(
-                  new ShiftedHoliday(
-                      new MonthDayHoliday(MonthDay.of(AUGUST, 11)), SUNDAY_TO_MONDAY),
-                  LocalDate.of(2020, 8, 11), // The Olympics
-                  LocalDate.of(2020, 8, 10),
-                  LocalDate.of(2021, 8, 11), // Postponement of the Olympics
-                  LocalDate.of(2021, 8, 9)),
-              ValueRange.of(2016, 2999)),
+                  new ShiftedHoliday(new MonthDayHoliday(AUGUST, 11), SUNDAY_TO_MONDAY),
+                  d(2020, 8, 11), // The Olympics
+                  d(2020, 8, 10),
+                  d(2021, 8, 11), // Postponement of the Olympics
+                  d(2021, 8, 9)),
+              2016,
+              2999),
           // https://wikipedia.org/wiki/Respect_for_the_Aged_Day - before 2003
           new YearRangeHoliday(
-              new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(SEPTEMBER, 15)), SUNDAY_TO_MONDAY),
-              ValueRange.of(1966, 2002)),
+              new ShiftedHoliday(new MonthDayHoliday(SEPTEMBER, 15), SUNDAY_TO_MONDAY), 1966, 2002),
           new BridgedHoliday(
               // https://wikipedia.org/wiki/Respect_for_the_Aged_Day - 2003 onward
               new YearRangeHoliday(
@@ -410,39 +400,40 @@ public enum FinancialCalendars implements Calendar {
           // https://wikipedia.org/wiki/Health_and_Sports_Day
           new MovedHoliday(
               new DayOfWeekInMonthHoliday(2, MONDAY, OCTOBER),
-              LocalDate.of(2020, 10, 12), // The Olympics
-              LocalDate.of(2020, 7, 24),
-              LocalDate.of(2021, 10, 11), // Postponement of the Olympics
-              LocalDate.of(2021, 7, 23)),
+              d(2020, 10, 12), // The Olympics
+              d(2020, 7, 24),
+              d(2021, 10, 11), // Postponement of the Olympics
+              d(2021, 7, 23)),
           // https://en.wikipedia.org/wiki/Culture_Day
-          new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(NOVEMBER, 3)), SUNDAY_TO_MONDAY),
+          new ShiftedHoliday(new MonthDayHoliday(NOVEMBER, 3), SUNDAY_TO_MONDAY),
           // https://www.timeanddate.com/holidays/japan/labor-thanksgiving-day
-          new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(NOVEMBER, 23)), SUNDAY_TO_MONDAY),
+          new ShiftedHoliday(new MonthDayHoliday(NOVEMBER, 23), SUNDAY_TO_MONDAY),
           // new years eve - bank of Japan, but not national holiday
-          new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(DECEMBER, 31)), SUNDAY_TO_MONDAY),
+          new ShiftedHoliday(new MonthDayHoliday(DECEMBER, 31), SUNDAY_TO_MONDAY),
           // https://en.wikipedia.org/wiki/The_Emperor%27s_Birthday - Akihito
           new YearRangeHoliday(
-              new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(DECEMBER, 23)), SUNDAY_TO_MONDAY),
-              ValueRange.of(1990, 2018)),
+              new ShiftedHoliday(new MonthDayHoliday(DECEMBER, 23), SUNDAY_TO_MONDAY), 1990, 2018),
           // https://www.timeanddate.com/holidays/japan/2019?hol=9
           new FixedHoliday(
-              LocalDate.of(
-                  2019, 4,
-                  30), // https://www.timeanddate.com/holidays/japan/coronation-day-holiday-1
-              LocalDate.of(2019, 5, 1), // https://www.timeanddate.com/holidays/japan/coronation-day
-              LocalDate.of(
-                  2019, 5,
-                  2), // https://www.timeanddate.com/holidays/japan/coronation-day-holiday-2
-              LocalDate.of(
-                  2019, 10,
-                  22) // https://www.timeanddate.com/holidays/japan/enthronement-ceremony-day
-              ),
+              // https://www.timeanddate.com/holidays/japan/coronation-day-holiday-1
+              d(2019, 4, 30),
+              // https://www.timeanddate.com/holidays/japan/coronation-day
+              d(2019, 5, 1),
+              // https://www.timeanddate.com/holidays/japan/coronation-day-holiday-2
+              d(2019, 5, 2),
+              // https://www.timeanddate.com/holidays/japan/enthronement-ceremony-day
+              d(2019, 10, 22)),
           // https://en.wikipedia.org/wiki/The_Emperor%27s_Birthday - Naruhito
           new YearRangeHoliday(
-              new ShiftedHoliday(new MonthDayHoliday(MonthDay.of(FEBRUARY, 23)), SUNDAY_TO_MONDAY),
-              ValueRange.of(2020, 2999))));
+              new ShiftedHoliday(new MonthDayHoliday(FEBRUARY, 23), SUNDAY_TO_MONDAY),
+              2020,
+              2999)));
 
   private final ConfigurableCalendar calendar;
+
+  private static LocalDate d(int year, int month, int day) {
+    return LocalDate.of(year, month, day);
+  }
 
   FinancialCalendars(ConfigurableCalendar calendar) {
     this.calendar = calendar;